src/pyproject.toml
[project]
name = "luckay.la"
version = "0.0.1"
description = "Personal site"
authors = [
    {name = "Aylawoo", email = "luckaylawoo@gmail.com"},
]
dependencies = [
    "jinja2>=3.1.4",
    "setuptools>=70.0.0",
    "python-dotenv>=1.0.1",
    "arrow>=1.3.0",
    "quart>=0.19.9",
]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
